+++ habit/src/main/java/app/habit/service/gpt/OpenAiService.java
@@ -1,16 +1,16 @@
 package app.habit.service.gpt;
 
 import app.habit.domain.Answer;
-import app.habit.domain.EvaluationGptCoach;
+import app.habit.service.gpt.coach.EvaluationGptCoach;
 import app.habit.domain.EvaluationPromptFactory;
-import app.habit.domain.FeedbackGptCoach;
+import app.habit.service.gpt.coach.FeedbackGptCoach;
 import app.habit.domain.FeedbackModule;
 import app.habit.domain.FeedbackPromptFactory;
 import app.habit.domain.FeedbackSession;
 import app.habit.domain.HabitAssessmentManager;
 import app.habit.domain.HabitFormingPhase;
-import app.habit.domain.PhaseGptCoach;
-import app.habit.domain.PreQuestionGptCoach;
+import app.habit.service.gpt.coach.PhaseGptCoach;
+import app.habit.service.gpt.coach.PreQuestionGptCoach;
 import app.habit.domain.PromptFactory;
 import app.habit.domain.Question;
 import app.habit.domain.SpecificPhasePreQuestionPromptFactory;
