+++ habit/src/main/java/app/habit/service/gpt/coach/FeedbackGptCoach.java
@@ -1,15 +1,13 @@
-package app.habit.domain;
+package app.habit.service.gpt.coach;
 
 import app.habit.dto.GptRsWrapper;
 import app.habit.dto.GptRsWrapper.Choice.Message;
-import app.habit.dto.UserHabitPreQuestionRs;
+import app.habit.dto.PhaseFeedbackRs;
 import app.habit.service.gpt.request.RequestPrompt;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -20,23 +18,20 @@ import org.springframework.web.client.RestTemplate;
 
 @Component
 @RequiredArgsConstructor
-public class PhaseGptCoach {
+public class FeedbackGptCoach {
 
     private static final String apiKey = "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";
 
     private final RestTemplate restTemplate;
 
-    public List<UserHabitPreQuestionRs> requestUserHabitPreQuestions(RequestPrompt requestBody, String url) {
-        GptRsWrapper adviceBody = writeAdvice(requestBody, url);
-        return parseAdvice(adviceBody);
+    public PhaseFeedbackRs requestPhaseFeedback(RequestPrompt requestBody, String url) {
+        HttpEntity<RequestPrompt> request = createAdviceRequest(requestBody);
+        GptRsWrapper advice = requestTotalAdvice(url, request);
+        return getCoreAdvice(advice);
     }
 
-    private GptRsWrapper writeAdvice(RequestPrompt requestBody, String url) {
-        return Optional.ofNullable(requestBody)
-                .map(body -> new HttpEntity<>(body, createHeaders()))
-                .map(entity -> restTemplate.exchange(url, HttpMethod.POST, entity, GptRsWrapper.class))
-                .map(HttpEntity::getBody)
-                .orElseThrow(NullPointerException::new);
+    private HttpEntity<RequestPrompt> createAdviceRequest(RequestPrompt requestBody) {
+        return new HttpEntity<>(requestBody, createHeaders());
     }
 
     private HttpHeaders createHeaders() {
@@ -47,10 +42,14 @@ public class PhaseGptCoach {
         return headers;
     }
 
-    private List<UserHabitPreQuestionRs> parseAdvice(GptRsWrapper body) {
+    private GptRsWrapper requestTotalAdvice(String url, HttpEntity<RequestPrompt> request) {
+        return restTemplate.exchange(url, HttpMethod.POST, request, GptRsWrapper.class).getBody();
+    }
+
+    private PhaseFeedbackRs getCoreAdvice(GptRsWrapper body) {
         Message message = body.getChoices().get(0).getMessage();
         String content = message.getContent();
-        TypeReference<List<UserHabitPreQuestionRs>> typeReference = new TypeReference<>() {
+        TypeReference<PhaseFeedbackRs> typeReference = new TypeReference<>() {
         };
         ObjectMapper objectMapper = new ObjectMapper();
         try {
